flow [nfc]: Move objectEntries to a more general flowPonyfill module
We'll soon give it a sibling.
